remove thermal changes and whitespace

+++ main/thermal/EMC2101.h
@@ -167,7 +167,7 @@ esp_err_t EMC2101_set_fan_speed(float);
 uint16_t EMC2101_get_fan_speed(void);
 esp_err_t EMC2101_init();
 float EMC2101_get_external_temp(void);
-float EMC2101_get_internal_temp(void);
+uint8_t EMC2101_get_internal_temp(void);
 esp_err_t EMC2101_set_ideality_factor(uint8_t);
 esp_err_t EMC2101_set_beta_compensation(uint8_t);
 #endif /* EMC2101_H_ */
